+++ R/model-lda.R
@@ -60,7 +60,8 @@ acceptable_formula.lda <- function(model) {
   acceptable_contrasts(
     columns = colnames(model$means),
     vars = names(attr(model$terms, "dataClasses")),
-    xlevels = model$xlevels
+    xlevels = model$xlevels,
+    terms = model$terms
   )
 
   acceptable_lm(model)